+++ test/Veggerby.Boards.Tests/Core/GameBuilderTests.cs
@@ -25,8 +25,8 @@ public class GameBuilderTests
             actual.State.Should().NotBeNull();
             actual.Engine.Should().NotBeNull();
 
-            actual.Game.Artifacts.Count().Should().Be(8);
-            actual.Game.Artifacts.Select(x => x.Id).Should().Equal(["piece-1", "piece-2", "piece-n", "piece-x", "piece-y", "dice", "dice-secondary", "artifact-x"]);
+            actual.Game.Artifacts.Count().Should().Be(9); // +1 TurnArtifact
+            actual.Game.Artifacts.Select(x => x.Id).Should().Equal(["piece-1", "piece-2", "piece-n", "piece-x", "piece-y", "dice", "dice-secondary", "artifact-x", "turn-timeline"]);
 
             actual.Game.Players.Count().Should().Be(2);
             actual.Game.Players.Select(x => x.Id).Should().Equal(["player-1", "player-2"]);
@@ -79,7 +79,7 @@ public class GameBuilderTests
             // assert
             actual.State.Should().NotBeNull();
             actual.State.IsInitialState.Should().BeTrue();
-            actual.State.ChildStates.Count().Should().Be(5);
+            actual.State.ChildStates.Count().Should().Be(6); // +1 TurnState
             actual.State.ChildStates.OfType<PieceState>().Count().Should().Be(3);
             actual.State.ChildStates.OfType<NullDiceState>().Count().Should().Be(1);
             actual.State.ChildStates.OfType<DiceState<int>>().Count().Should().Be(1);